GOVSI-1166 - Move ResponseHeaders to shared
- As they are used in multiple places now and to prevent duplication
