--- oidc-api/src/main/java/uk/gov/di/authentication/oidc/entity/ResponseHeaders.java
@@ -1,6 +1,0 @@
-package uk.gov.di.authentication.oidc.entity;
-
-public interface ResponseHeaders {
-    String LOCATION = "Location";
-    String SET_COOKIE = "Set-Cookie";
-}
